% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Plot a barplot of the given values and highlight one value}
\usage{
bar_plot(values, to_highlight, colors = customColors1,
  plotTheme = theme_classic())
}
\arguments{
\item{values}{numeric vector to plot}

\item{to_highlight}{position to highlight}
}
\value{
ggplot2 plot object
}
\description{
Plot a barplot of the given values and highlight one value
}
