% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthbarometer_value.R
\name{healthbarometer_value}
\alias{healthbarometer_value}
\title{Caclulate mean of indices for overall healthbaromter value}
\usage{
healthbarometer_value(values, user_weights)
}
\arguments{
\item{values}{dataframe: containing the all relevant variables}

\item{weigths}{numeric vector: containnig the weight for each variabel in df}
}
\value{
numeric vector: value for healthbarometer
}
\description{
Caclulate mean of indices for overall healthbaromter value
}
