% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{Add a placeholder to a shiny page}
\usage{
placeholder(size)
}
\arguments{
\item{size}{no of line breaks to be inserted}
}
\description{
Add a placeholder to a shiny page
}
