% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_raw_data.R
\name{transform_raw_data}
\alias{transform_raw_data}
\title{Transform and save raw data in form of an r object}
\usage{
transform_raw_data(path = "data/data_indicators.csv", save = TRUE)
}
\arguments{
\item{string}{scalar: path to raw data as csv}

\item{boolean}{sclar: whether data.rds in data should be rewritten?}
}
\value{
data frame: converted data from source
}
\description{
Transform and save raw data in form of an r object
}
