% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_percentilelike.R
\name{transform_to_percentilelike}
\alias{transform_to_percentilelike}
\title{Transform variable to percentile depiction}
\usage{
transform_to_percentilelike(var)
}
\arguments{
\item{var}{numeric scalar: variable to be transformed}
}
\value{
numeric scalar: transformed variable
}
\description{
Transform variable to percentile depiction
}
